backend/pyproject.toml
[project]
name = "bolus-ai-backend"
version = "0.1.0"
description = "Bolus AI FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.29.0",
    "httpx==0.27.0",
    "pydantic==1.10.15",
    "python-jose==3.3.0",
    "passlib[bcrypt]==1.7.4",
]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
